Add minigrant application to resources (#21)
* Isla's first blog post

* Added author line

* updated authors

* no really updated authors this time

* Update authors.yml

* Update authors.yml

* Add files via upload

* Update 2017-05-12-Hello-Earthling.md

* add Brain Networks In Python blog post

* Changed silly name

* edit blog post

* add minigrant application to resources

* add jekyll front matter to minigrant resource

* remove blog from branch

* Adds Isla as author to blog

This commit includes Kirstie's edits to the mozilla mini grant
resource post. It renamed the file so it has the correct naming
style, updates the date (because it's taken Kirstie a while
to actually review this pull!) and adds Isla as the author of
the post.

* Delete images associated with blog post

These images were left over from a different branch, so I'm
deleting them here. They can absolutely be added back in with the
blog post!

* Fixes 401 error for link to Bullmore & Sporns paper

* Fix (again) link to Bullmore & Sporns paper

This is really annoying - the link to the Nature Reviews
Neuroscience page is giving a 401 error which means the travis
build is failing :sad: Fingers crossed this new link to pubmed
isn't a problem.