Interesting papers blog post (#23)
* Add blog post with useful papers

This blog post includes a bunch of papers that I've been meaning to email
out to the Bullmore lab group.

I figured I'd write it up as a blog post instead incase it was useful to
others!

I've also added in an author field to the blog template to remind others
to make sure it doesn't default to me in the future.

* Fix 401 link after travis build fail

This really annoys me but because the paper is behind a firewall
Travis says that the link is "broken". I really want to have a
beautiful green button for the travis build so I've replaced the
journal link with the PubMed one.