Blog: Animation Sspotlight and Blog Article Draft
Hello Santiago,

I've done the draft for the Animation Film Spotlight and the Blog Article Section. I'm struggling with the category name for the articles that will be helpful for professionals, hobbyist and students alike. At first I thought of Monthly Musings, but it sounded too wishy washy, then I changed to Monthly Digest. I've considered "CG Insider" but it sounded like a completely new brand, or 3D media wannabe. The current one that I just thought of is Essential Tips (continuing from the same direction as "tips and tricks" or 3D Essentials. Or maybe, as you have suggested, Tips and Tricks will work just fine because it's broad enough. Maybe we can go over this tomorrow?